Remove unused types.
